Pagination

+++ database/migrations/2016_10_23_204233_create_vehicles_table.php
@@ -17,8 +17,8 @@ class CreateVehiclesTable extends Migration
             $table->increments('serie');
             $table->string('color');
             $table->integer('power');
-            $table->float('capacity');
-            $table->float('speed');
+            $table->double('capacity');
+            $table->double('speed');
             $table->integer('maker_id')->unsigned()->default(0);
             $table->foreign('maker_id')->references('id')->on('makers')->onDelete('cascade');
             $table->timestamps();
